backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.36.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.17
